fix log level criteria holder

+++ src/main/java/com/epam/ta/reportportal/commons/querygen/CriteriaHolder.java
@@ -1,3 +1,19 @@
+/*
+ *  Copyright (C) 2018 EPAM Systems
+ *
+ *  Licensed under the Apache License, Version 2.0 (the "License");
+ *  you may not use this file except in compliance with the License.
+ *  You may obtain a copy of the License at
+ *
+ *  http://www.apache.org/licenses/LICENSE-2.0
+ *
+ *  Unless required by applicable law or agreed to in writing, software
+ *  distributed under the License is distributed on an "AS IS" BASIS,
+ *  WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ *  See the License for the specific language governing permissions and
+ *  limitations under the License.
+ */
+
 package com.epam.ta.reportportal.commons.querygen;
 
 import com.epam.ta.reportportal.commons.validation.BusinessRule;
@@ -102,7 +118,7 @@ public class CriteriaHolder {
 		} else if (boolean.class.equals(getDataType()) || Boolean.class.isAssignableFrom(getDataType())) {
 			castedValue = BooleanUtils.toBoolean(oneValue);
 		} else if (LogLevel.class.isAssignableFrom(getDataType())) {
-			castedValue = LogLevel.toLevel(oneValue);
+			castedValue = LogLevel.toLevel(oneValue).toInt();
 			BusinessRule.expect(castedValue, Objects::nonNull)
 					.verify(errorType, Suppliers.formattedSupplier("Cannot convert '{}' to valid 'LogLevel'", oneValue));
 		} else if (TestItemIssueGroup.class.isAssignableFrom(getDataType())) {
